requirements.txt
aiogram
python-dotenv
sqlalchemy
aiosqlite
alembic
pydantic-settings
greenlet
requests
pytest-asyncio
coverage
google-auth-oauthlib
